+++ assets/js/disable-search.js
@@ -1,0 +1,3 @@
+wp.domReady( () => {
+	wp.blocks.unregisterBlockType( 'core/search' );
+} );